feat: Add support for Task returning methods
In the previous version, when a return value of a method was a Task, a
continuation was set on that task to execute the After/Catch/Finally
flow, but that continuation was independent from the overall method
invocation.

That means that the continuation would run after the method
has already returned and unlocked any caller waiting for the resulting
task.

These changes will make the After/Catch/Finally block run within the
scope of the call.